+++ world/src/layer/spot.rs
@@ -64,7 +64,7 @@ pub enum Spot {
     FruitTree,
     Shipwreck,
     Shipwreck2,
-    GnarlingTree,
+    FallenTree,
     TrollCave,
     TrollCaveMountain,
     TrollCaveSwamp,
@@ -220,7 +220,7 @@ impl Spot {
         Self::generate_spots(
             Spot::GnarlingTotem,
             world,
-            2.0,
+            1.5,
             |g, c| {
                 g < 0.25
                     && !c.near_cliffs()
@@ -232,9 +232,9 @@ impl Spot {
             false,
         );
         Self::generate_spots(
-            Spot::GnarlingTree,
+            Spot::FallenTree,
             world,
-            1.0,
+            1.5,
             |g, c| {
                 g < 0.25
                     && !c.near_cliffs()
@@ -601,9 +601,9 @@ pub fn apply_spots_to(canvas: &mut Canvas, _dynamic_rng: &mut impl Rng) {
                 base_structures: Some("spots.myrmidon-temple"),
                 entity_radius: 10.0,
                 entities: &[
-                    (3..5, "common.entity.dungeon.tier-4.hoplite"),
-                    (3..5, "common.entity.dungeon.tier-4.strategian"),
-                    (2..3, "common.entity.dungeon.tier-4.marksman"),
+                    (3..5, "common.entity.dungeon.myrmidon.hoplite"),
+                    (3..5, "common.entity.dungeon.myrmidon.strategian"),
+                    (2..3, "common.entity.dungeon.myrmidon.marksman"),
                 ],
             },
             Spot::WitchHouse => SpotConfig {
@@ -645,14 +645,13 @@ pub fn apply_spots_to(canvas: &mut Canvas, _dynamic_rng: &mut impl Rng) {
                     (1..2, "common.entity.dungeon.gnarling.woodgolem"),
                 ],
             },
-            Spot::GnarlingTree => SpotConfig {
-                base_structures: Some("spots_grasslands.gnarling_tree"),
+            Spot::FallenTree => SpotConfig {
+                base_structures: Some("spots_grasslands.fallen_tree"),
                 entity_radius: 64.0,
                 entities: &[
-                    (1..5, "common.entity.dungeon.gnarling.mugger"),
-                    (2..4, "common.entity.dungeon.gnarling.stalker"),
-                    (1..2, "common.entity.dungeon.gnarling.logger"),
-                    (1..4, "common.entity.wild.aggressive.deadwood"),
+                    (1..2, "common.entity.dungeon.gnarling.mandragora"),
+                    (2..6, "common.entity.wild.aggressive.deadwood"),
+                    (0..1, "common.entity.wild.aggressive.mossdrake"),
                 ],
             },
             Spot::TrollCave => SpotConfig {